code/backend/requirements.txt
flask
flask-cors
mysql-connector-python  # This is the package you're using for MySQL connections
python-dotenv
